+++ shiny/templates/chat/hello-providers/ollama/app.py
@@ -5,8 +5,9 @@
 # To install the Ollama Python client, see https://github.com/ollama/ollama-python
 # ------------------------------------------------------------------------------------
 
+from chatlas import Ollama
+
 from shiny.express import ui
-from shiny.ui._chat_client_ollama import OllamaClient
 
 # Set some Shiny page options
 ui.page_opts(
@@ -16,7 +17,7 @@ ui.page_opts(
 )
 
 # Assumes you're running an Ollama server (with llama3 available) locally
-llm = OllamaClient(model="llama3.1")
+llm = Ollama(model="llama3.1")
 
 # Create and display empty chat
 chat = ui.Chat(
@@ -29,5 +30,5 @@ chat.ui()
 # Define a callback to run when the user submits a message
 @chat.on_user_submit
 async def _(input):
-    response = llm.generate_response(input)
+    response = llm.response_generator(input)
     await chat.append_message_stream(response)