+++ shiny/templates/chat/hello-providers/openai/app.py
@@ -1,9 +1,9 @@
 import os
 
 from app_utils import load_dotenv
+from chatlas import OpenAI
 
 from shiny.express import ui
-from shiny.ui._chat_client_openai import OpenAIClient
 
 ui.page_opts(
     title="Hello OpenAI Chat",
@@ -12,7 +12,7 @@ ui.page_opts(
 )
 
 load_dotenv()
-llm = OpenAIClient(
+llm = OpenAI(
     api_key=os.environ.get("OPENAI_API_KEY"),
     model="gpt-4o",
 )
@@ -27,5 +27,5 @@ chat.ui()
 
 @chat.on_user_submit
 async def _(input):
-    response = llm.generate_response(input, stream=False)
-    await chat.append_message(response)
+    response = llm.response_generator(input)
+    await chat.append_message_stream(response)